Regularize the input buttons.
The buttons are now labeled 'preview' 'submit attempt' and 'grade and finish'.  Previewing shows you type set versions of the answers entered and doesn't count as as "try".  Submit counts as a try and displays the correctness of each entry (if showPartialAnswers  = 1).  'Grade and Finish' enters the grade into the database and disables all of the buttons -- you can't enter any data after the grade and finish button is pressed.

This is pretty close to standard webwork homework assignment behavior.  I was trying to see how close I could come within the moodle quiz framework.
It can probably be improved further but this gives everyone something to work with.